+++ client/src/pages/Game/pages/Table/components/InYourFaceEvents/InYourFaceEvents.tsx
@@ -1,6 +1,7 @@
 import { Button, Portal } from '@/components'
 import { ClippedBox } from '@/components/ClippedBox'
 import { Flex } from '@/components/Flex/Flex'
+import { useLocale } from '@/context/LocaleContext'
 import { useAppStore, useToggle } from '@/utils/hooks'
 import { useGameEventsHandler } from '@/utils/useGameEventsHandler'
 import { faBell, faChevronDown } from '@fortawesome/free-solid-svg-icons'
@@ -8,43 +9,10 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { EventType, GameEvent } from '@shared/index'
 import { useCallback, useEffect, useState } from 'react'
 import styled, { css, keyframes } from 'styled-components'
-import { CardPlayedEvent } from './components/CardPlayedEvent'
-import { CardUsedEvent } from './components/CardUsedEvent'
-import { ColonyBuiltEvent } from './components/ColonyBuiltEvent'
-import { ColonyTradingEvent } from './components/ColonyTradingEvent'
-import { CompetitionSponsoredEvent } from './components/CompetitionSponsoredEvent'
-import { CurrentGlobalEventExecutedEvent } from './components/CurrentGlobalEventExecuted'
-import { GlobalEventsChangedEvent } from './components/GlobalEventsChangedEvent'
-import { MarsTerraformedEvent } from './components/MarsTerraformedEvent'
-import { MilestoneBoughtEvent } from './components/MilestoneBoughtEvent'
-import { NewGovernmentEvent } from './components/NewGovernmentEvent'
+import { InYourFaceEvent } from './components/InYourFaceEvent'
 import { PlayerDidHeader } from './components/PlayerDidHeader'
-import { PlayerMovedDelegateEvent } from './components/PlayerMovedDelegateEvent'
-import { ProductionDoneEvent } from './components/ProductionDoneEvent'
-import { StandardProjectBoughtEvent } from './components/StandardProjectBoughtEvent'
-import { StartingSetupEvent } from './components/StartingSetupEvent'
-import { TilePlacedEvent } from './components/TilePlacedEvent'
-import { CommitteePartyActivePolicyActivatedEvent } from './components/CommitteePartyActivePolicyActivatedEvent'
-import { useLocale } from '@/context/LocaleContext'
-
-const PROCESSABLE_EVENTS = [
-	EventType.CardPlayed,
-	EventType.CardUsed,
-	EventType.StandardProjectBought,
-	EventType.MilestoneBought,
-	EventType.CompetitionSponsored,
-	EventType.ColonyBuilt,
-	EventType.ColonyTrading,
-	EventType.StartingSetup,
-	EventType.ProductionDone,
-	EventType.TilePlaced,
-	EventType.MarsTerraformed,
-	EventType.GlobalEventsChanged,
-	EventType.CurrentGlobalEventExecuted,
-	EventType.NewGovernment,
-	EventType.PlayerMovedDelegate,
-	EventType.CommitteePartyActivePolicyActivated,
-]
+import { isInYourFaceEvent } from './utils/isInYourFaceEvent'
+import { InYourFaceEventsList } from './components/InYourFaceEventsList'
 
 export const InYourFaceEvents = () => {
 	const player = useAppStore((state) => state.game.player)
@@ -56,6 +24,8 @@ export const InYourFaceEvents = () => {
 	const [shown, toggleShown, setShown] = useToggle()
 	const [rendered, setRendered] = useState(false)
 	const [events, setEvents] = useState<GameEvent[]>([])
+	const [showList, toggleList] = useToggle()
+
 	const t = useLocale()
 
 	const current = events[0]
@@ -65,11 +35,7 @@ export const InYourFaceEvents = () => {
 			return
 		}
 
-		if (event.processed) {
-			return
-		}
-
-		if (PROCESSABLE_EVENTS.includes(event.type)) {
+		if (isInYourFaceEvent({ event })) {
 			setEvents((events) => [...events, event])
 		}
 	})
@@ -189,48 +155,17 @@ export const InYourFaceEvents = () => {
 		}
 	}, [])
 
-	const renderEvent = useCallback((event: GameEvent) => {
-		switch (event.type) {
-			case EventType.CardPlayed:
-				return <CardPlayedEvent event={event} />
-			case EventType.CardUsed:
-				return <CardUsedEvent event={event} />
-			case EventType.StandardProjectBought:
-				return <StandardProjectBoughtEvent event={event} />
-			case EventType.MilestoneBought:
-				return <MilestoneBoughtEvent event={event} />
-			case EventType.CompetitionSponsored:
-				return <CompetitionSponsoredEvent event={event} />
-			case EventType.ColonyBuilt:
-				return <ColonyBuiltEvent event={event} />
-			case EventType.ColonyTrading:
-				return <ColonyTradingEvent event={event} />
-			case EventType.StartingSetup:
-				return <StartingSetupEvent event={event} />
-			case EventType.ProductionDone:
-				return <ProductionDoneEvent event={event} />
-			case EventType.TilePlaced:
-				return <TilePlacedEvent event={event} />
-			case EventType.MarsTerraformed:
-				return <MarsTerraformedEvent />
-			case EventType.GlobalEventsChanged:
-				return <GlobalEventsChangedEvent event={event} />
-			case EventType.CurrentGlobalEventExecuted:
-				return <CurrentGlobalEventExecutedEvent event={event} />
-			case EventType.NewGovernment:
-				return <NewGovernmentEvent event={event} />
-			case EventType.PlayerMovedDelegate:
-				return <PlayerMovedDelegateEvent event={event} />
-			case EventType.CommitteePartyActivePolicyActivated:
-				return <CommitteePartyActivePolicyActivatedEvent event={event} />
-			default:
-				return null
-		}
-	}, [])
-
 	const handleDismiss = useCallback(() => {
 		setEvents((events) => events.slice(1))
 	}, [])
+
+	const handleShow = () => {
+		if (current) {
+			toggleShown()
+		} else {
+			toggleList()
+		}
+	}
 
 	useEffect(() => {
 		setShown(true)
@@ -238,12 +173,14 @@ export const InYourFaceEvents = () => {
 
 	return (
 		<>
-			<MinimizedButton onClick={toggleShown} noClip>
+			<MinimizedButton onClick={handleShow} noClip>
 				<ButtonIcon>
 					<FontAwesomeIcon icon={faBell} />
 				</ButtonIcon>
 				<ButtonCount>{events.length}</ButtonCount>
 			</MinimizedButton>
+			{showList && <InYourFaceEventsList onClose={toggleList} />}
+
 			{current && rendered && (
 				<Portal>
 					<DisplayContainer
@@ -291,7 +228,7 @@ export const InYourFaceEvents = () => {
 							</Actions>
 							{/* TODO: Hotfix to reset animations for new events even when they're the same type - better solution? */}
 							<Event key={JSON.stringify(current)}>
-								{renderEvent(current)}
+								<InYourFaceEvent event={current} />
 							</Event>
 							<div></div>
 						</Inner>
